+++ dmf.h
@@ -2,25 +2,19 @@
 dmf.h
 Written by Dalton Messmer <messmer.dalton@gmail.com>. 
 
-Provides functions for loading a .dmf file according to the 
+Provides functions for loading a dmf file according to the 
 spec sheet at http://www.deflemask.com/DMF_SPECS.txt. 
 
 Requires the zlib compression library from https://zlib.net. 
 */
 
-#ifndef __DMF_H__
-#define __DMF_H__
+#pragma once
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <stdint.h>
 #include <assert.h>
-#include <stdbool.h>
-
-// For inflating .dmf files so that they can be read 
-#include "zlib.h"
-#include "zconf.h"
 
 // Deflemask allows four effects columns per channel regardless of the system 
 #define MAX_EFFECTS_COLUMN_COUNT 4 
@@ -65,23 +59,23 @@ typedef enum DMF_GAMEBOY_EFFECT {
 
 typedef struct Note 
 {
-    uint16_t pitch; 
-    uint16_t octave;  
-} Note; 
+    uint16_t pitch;
+    uint16_t octave;
+} Note;
 
 typedef struct System
 {
     uint8_t id;
-    char *name;
+    const char* name;
     uint8_t channels;
 } System;
 
-// SYSTEM_TYPE values also correspond to indices in Systems array. 
-typedef enum SYSTEM_TYPE 
+// SYSTEM_TYPE values also correspond to indices in DMF::m_Systems array.
+typedef enum SYSTEM_TYPE
 {
-    SYS_ERROR, SYS_GENESIS, SYS_GENESIS_CH3, SYS_SMS, SYS_GAMEBOY, 
+    SYS_ERROR=0, SYS_GENESIS, SYS_GENESIS_CH3, SYS_SMS, SYS_GAMEBOY, 
     SYS_PCENGINE, SYS_NES, SYS_C64_SID_8580, SYS_C64_SID_6581, SYS_YM2151
-} SYSTEM_TYPE; 
+} SYSTEM_TYPE;
 
 typedef struct VisualInfo
 {
@@ -143,42 +137,87 @@ typedef struct PatternRow
     int16_t instrument;
 } PatternRow; 
 
-typedef struct DMFContents
+typedef enum DMF_IMPORT_ERROR
 {
-    uint8_t dmfFileVersion; 
-    System sys; 
-    VisualInfo visualInfo; 
-    ModuleInfo moduleInfo; 
-    uint8_t **patternMatrixValues; 
-    uint8_t *patternMatrixMaxValues; 
-    uint8_t totalInstruments; 
-    Instrument* instruments; 
-    uint8_t totalWavetables; 
-    uint32_t *wavetableSizes;   
-    uint32_t **wavetableValues; 
-    PatternRow ***patternValues; 
-    uint8_t *channelEffectsColumnsCount; 
-    uint8_t totalPCMSamples; 
-    PCMSample *pcmSamples; 
-} DMFContents; 
+    IMPORT_ERROR_SUCCESS=0,
+    IMPORT_ERROR_FAIL=1
+} DMF_IMPORT_ERROR;
+
+
+
+class DMF
+{
+public:
+    DMF(const char* filename);
+    ~DMF();
+
+    static System SYSTEMS(SYSTEM_TYPE systemType) { return m_Systems[systemType]; }
+
+    DMF_IMPORT_ERROR Status() { return m_ImportError; };
+
+    // Returns the initial BPM of the module when given ModuleInfo
+    double GetBPM();
+
+    const System& GetSystem() const { return m_System; }
+    const VisualInfo& GetVisualInfo() const { return m_VisualInfo; }
+    const ModuleInfo& GetModuleInfo() const { return m_ModuleInfo; }
+
+    uint8_t** const GetPatternMatrixValues() const { return m_PatternMatrixValues; }
+
+    uint8_t GetTotalWavetables() const { return m_TotalWavetables; }
+
+    uint32_t** const GetWavetableValues() const { return m_WavetableValues; }
+    uint32_t GetWavetableValue(unsigned wavetable, unsigned index) { return m_WavetableValues[wavetable][index]; }
+
+    PatternRow*** const GetPatternValues() { return m_PatternValues; }
+
+private:
+    System GetSystem(uint8_t systemByte);
+    void LoadVisualInfo(uint8_t **fBuff, uint32_t *pos);
+    void LoadModuleInfo(uint8_t **fBuff, uint32_t *pos);
+    void LoadPatternMatrixValues(uint8_t **fBuff, uint32_t *pos);
+    void LoadInstrumentsData(uint8_t **fBuff, uint32_t *pos);
+    Instrument LoadInstrument(uint8_t **fBuff, uint32_t *pos, System systemType);
+    void LoadWavetablesData(uint8_t **fBuff, uint32_t *pos);
+    void LoadPatternsData(uint8_t **fBuff, uint32_t *pos);
+    PatternRow LoadPatternRow(uint8_t **fBuff, uint32_t *pos, int effectsColumnsCount);
+    void LoadPCMSamplesData(uint8_t **fBuff, uint32_t *pos);
+    PCMSample LoadPCMSample(uint8_t **fBuff, uint32_t *pos);
+
+private:
+
+    static const System m_Systems[];
+
+    uint8_t         m_DMFFileVersion;
+    System          m_System;
+    VisualInfo      m_VisualInfo;
+    ModuleInfo      m_ModuleInfo;
+    uint8_t**       m_PatternMatrixValues;
+    uint8_t*        m_PatternMatrixMaxValues;
+    uint8_t         m_TotalInstruments;
+    Instrument*     m_Instruments;
+    uint8_t         m_TotalWavetables;
+    uint32_t*       m_WavetableSizes;
+    uint32_t**      m_WavetableValues;
+    PatternRow***   m_PatternValues;
+    uint8_t*        m_ChannelEffectsColumnsCount;
+    uint8_t         m_TotalPCMSamples;
+    PCMSample*      m_PCMSamples;
+
+    DMF_IMPORT_ERROR m_ImportError;
+};
 
 // Deflemask Game Boy channels 
 typedef enum DMF_GAMEBOY_CHANNEL {
     DMF_GAMEBOY_SQW1=0, DMF_GAMEBOY_SQW2=1, DMF_GAMEBOY_WAVE=2, DMF_GAMEBOY_NOISE=3
 } DMF_GAMEBOY_CHANNEL; 
 
-// Imports the .dmf file "fname" and stores it in the struct "dmf" 
-int importDMF(const char *fname, DMFContents *dmf); 
-
-// Returns the initial bpm of the module when given ModuleInfo 
-double getBPM(const ModuleInfo *info);  
+// Imports the .dmf file "fname" and stores it in the dmf object
+int importDMF(const char *fname, DMF *dmf); 
 
 // Compares notes n1 and n2. Returns 1 if n1 > n2, -1 if n1 < n2, and 0 if n1 == n2. 
-int8_t noteCompare(const Note *n1, const Note *n2); 
+int8_t NoteCompare(const Note *n1, const Note *n2);
 
-// Frees the dynamically allocated memory used by a DMFContents struct 
-void freeDMF(DMFContents *dmf); 
+int8_t NoteCompare(const Note* n1, const Note n2);
 
-char *getFilenameExt(const char *fname); 
-
-#endif 
+const char* GetFilenameExt(const char *fname);
